Adding a small test environment based around the unittest framework.
Alongside it, I have added 'requirements.txt' and 'test-requirements.txt' so that we can start defining dependencies for our project. These dependencies are still not referenced from the 'setup.py' file, but they are on the test environment. Additionally, I have added a dummy test file so that we can see the test environment working.