Create environments based on loaded configuration
Once the app has loaded the configuration, it has all
the data required for creating the different
Environment instances.

These environments will be later populated with information
gathered from the different sources.

In order for the app to work properly and for tests to pass,
this change adds also an empty Source class.

Finally, added InvalidConfiguration exception to be raised
whenever an invalid configuration was set by the user. (e.g. missing
system information)