Extend parser with arguments from CI models
In summary, initial parser is updated with additional
arguments from the different CI models created after
the configuration is loaded.

Each argument belongs to its own parser group which determined
by the class of the object it belongs to. Basically, each CI model
has its own group in the parser. This way, argument clearly separated
into different groups and more easily identified by the user.

Part of this change, some information on the CI models was moved
from instance scope to class scope. This is because the way arguments
are obtained is from the relationship between the different CI model
classes. If it was based on actual instances, we couldn't get the
information from the all CI models because when the user first runs
the application, we don't actually create instances for job.